refactor: split compoents and fix style show err
